+++ packages/pinia/src/storeToRefs.ts
@@ -3,12 +3,12 @@ import {
   isReactive,
   isRef,
   isVue2,
-  Ref,
   toRaw,
   ToRef,
   toRef,
   ToRefs,
   toRefs,
+  WritableComputedRef,
 } from 'vue-demi'
 import { StoreGetters, StoreState } from './store'
 import type {
@@ -21,8 +21,29 @@ import type {
   StoreGeneric,
 } from './types'
 
-type ToComputedRefs<T> = {
-  [K in keyof T]: ToRef<T[K]> extends Ref ? ComputedRef<T[K]> : ToRef<T[K]>
+/**
+ * Internal utility type
+ */
+type _IfEquals<X, Y, A = true, B = false> =
+  (<T>() => T extends X ? 1 : 2) extends <T>() => T extends Y ? 1 : 2 ? A : B
+
+/**
+ * Internal utility type
+ */
+type _IsReadonly<T, K extends keyof T> = _IfEquals<
+  { [P in K]: T[P] },
+  { -readonly [P in K]: T[P] },
+  false, // Property is not readonly if they are the same
+  true // Property is readonly if they differ
+>
+
+/**
+ * Extracts the getters of a store while keeping writable and readonly properties. **Internal type DO NOT USE**.
+ */
+type _ToComputedRefs<SS> = {
+  [K in keyof SS]: true extends _IsReadonly<SS, K>
+    ? ComputedRef<SS[K]>
+    : WritableComputedRef<SS[K]>
 }
 
 /**
@@ -47,9 +68,14 @@ type _ToStateRefs<SS> =
  * Extracts the return type for `storeToRefs`.
  * Will convert any `getters` into `ComputedRef`.
  */
-export type StoreToRefs<SS extends StoreGeneric> = _ToStateRefs<SS> &
-  ToRefs<PiniaCustomStateProperties<StoreState<SS>>> &
-  ToComputedRefs<StoreGetters<SS>>
+export type StoreToRefs<SS extends StoreGeneric> =
+  // NOTE: always trues but the conditional makes the type distributive
+  // https://www.typescriptlang.org/docs/handbook/2/conditional-types.html#distributive-conditional-types
+  SS extends unknown
+    ? _ToStateRefs<SS> &
+        ToRefs<PiniaCustomStateProperties<StoreState<SS>>> &
+        _ToComputedRefs<StoreGetters<SS>>
+    : never
 
 /**
  * Creates an object of references with all the state, getters, and plugin-added